perf: move docs layout into catchall page
